Fix Python 3 compatibility
This code include a couple of uses of iteritems(), which no longer
exists in Python 3.  items() will work for both, but the behavior is
technically different.  The performance will be slightly worse with
Python 2 since it will no longer be using an iterator, but everyone
should be using Python 3 at this point, anyway.  items() returns an
iterator with Python 3.